clean up mint
