+++ x/mint/abci_app.go
@@ -2,10 +2,11 @@ package mint
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/mint/keeper"
 )
 
 // BeginBlocker inflates every block and updates inflation parameters once per hour
-func BeginBlocker(ctx sdk.Context, k Keeper) {
+func BeginBlocker(ctx sdk.Context, k keeper.Keeper) {
 
 	// fetch stored minter & params
 	minter := k.GetMinter(ctx)
