+++ x/mint/keeper/expected_keepers.go
@@ -1,4 +1,4 @@
-package mint
+package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
