+++ x/mint/keeper/querier.go
@@ -1,4 +1,4 @@
-package mint
+package keeper
 
 import (
 	"fmt"
