+++ x/mint/keeper/test_common.go
@@ -1,4 +1,4 @@
-package mint
+package keeper
 
 import (
 	"os"
