+++ x/mint/types/params.go
@@ -1,4 +1,4 @@
-package mint
+package types
 
 import (
 	"fmt"
